chore(remap): re-work cli
